requirements.txt
Django
djangorestframework
mysqlclient
requests

